+++ apps/rest-api-server/src/clients/prisma.ts
@@ -18,6 +18,9 @@ export const prisma = getPrisma({
         apiBaseUrl: env.S3_STORAGE_ENDPOINT,
         bucketName: env.S3_STORAGE_BUCKET_NAME,
       },
+      ipfs: {
+        gatewayUrl: env.IPFS_STORAGE_GATEWAY_URL,
+      },
     },
   },
 });