+++ apps/web/src/components/ControlBars/ListControlBar.tsx
@@ -95,10 +95,10 @@ export const ListControlBar: FC = () => {
                 inputEndProps={{
                   placeholder: `End ${RANGE_LABEL[rangeType as keyof typeof RANGE_LABEL]}`,
                 }}
-                range={state.range.values}
+                range={state.range.values as { start: number | null; end: number | null }}
                 onChange={({ end = null, start = null }) =>
                   actions.setRange({
-                    type: rangeType,
+                    type: rangeType as "block" | "slot" | "epoch",
                     values: {
                       end,
                       start,